only set fetch's credentials:include option if no access token (#234)
* only set fetch's credentials:include option if no access token

this is what client.js does, and fixes a CORS issue when running client4.js in a browser.
Specifically, the issue seen is:

Fetch API cannot load http://127.0.0.1:8065/api/v4/users/me. Response to preflight request doesn't pass access control
check: The value of the 'Access-Control-Allow-Credentials' header in the response is '' which must be 'true' when the
request's credentials mode is 'include'. Origin 'http://127.0.0.1:8002' is therefore not allowed access.

* Revert "only set fetch's credentials:include option if no access token"

This reverts commit 2ba1306a69e46f84a7143c2ce2753a1f222cb118.

* make sending credentials via fetch configurable in client4.js

this preserves the default behaviour of sending cookies as part of the request, but can
be turned off with client.setIncludeCookes(false).