ABC-216: fix tutorial progression (#716)
* ABC-216: fix tutorial progression

This change fixes a `===` comparison between a string an integer to
allow the second tutorial step to display. It also fixes the trigger
that increments the tutorial step to only fire on an explicit "next"
instead of twice on a click because of multiple registered event
handlers. This also means that the tutorial no longer advances on an
"escape", but requires being explicitly skipped or completed.

As part of the clean up, unify most handling of the tutorial step
preference constants.

* ABC-234: fix clicking on tutorial circles

The circles were effectively setting a screen state of `"1"` instead of
the expected `1`.

* ABC-216: switch to arrow functions