Remove mm_config and mm_license global state from webapp (PR #1) (#816)
* MM-8589: change i18n/getLanguages to use store

This also removes the caching of the available languages data structure,
as it obviously needs to be able to change. It's not on a frequently
used code path so the caching appears to (at least now) be unnecessary.

* MM-8589: eliminate UserStore.getNoAccounts()

No changes to the use of global variables here. Rather, I'm just pushing
down an indirect use of same.

* MM-8589: expunge global mm_(config|license) from claim components

This also eliminates `LdapPasswordFieldName` as the server-side
configuration does not exist.

* MM-8589: clarify props passed to claim components

* MM-8589: expunge unnecessary ownProps

* MM-8589: use the get(Config|License) selector