MM-8604: process config/license websocket events (#418)
* MM-8604: listen for websocket events to update license and config

Rework the existing `CONFIG_CHANGED` event into `SERVER_VERSION_CHANGED`
and rebrand `CONFIG_CHANGED` to be emitted only on configuration
changes.

* remove broken handling of CONFIG_CHANGED event

This fired an action creator, but didn't do anything with the return
value so no action was ever dispatched to the store. The mobile
repository handles this event itself, and the webapp doesn't currently
require it.