MM-26053: change DM name format according to teammate name display se… (#6095)
* MM-26053: change DM name format according to teammate name display setting

When the preference is set to display full_name, we display the full name first,
and then the username. This helps keep the correct order in between GMs and DMs
while searching.

This is because the GM display already changes according to that setting,
but the DM name did not. So that caused confusion when searching for a full name
because giving a space after the first name would cause the GM to jump ahead of the DM.

This keeps things consistent.

* Make unit tests have separate states

Co-authored-by: Mattermod <mattermod@users.noreply.github.com>